+++ CHANGELOG.md
@@ -40,6 +40,7 @@
  - Added event listener to Environment Service (#1996)
  - Autostarted agents can load a new value for the autostart_agent_map setting without agent restart (#1839)
  - Added protected environment option (#1997)
+ - Added warning when trying to override a built-in type with a typedef (#81)
 
 # v 2020.1 (2020-02-19) Changes in this release: